Add new `Signal` generators including: - Sine - Saw - Square - Noise - NoiseSimplex
Add a `bus` method to the `Signal` trait which allows for outputting a
`Signal` to multiple other signals. Still unsure about current method as
there needs to be some way of restricting the rate of output signals to
the same rate and ensuring they are called simultaneously, as if they
are called out of sync pending samples need to be stored which may
result in large leakage.

Changed FromSamples to FromInterleavedSamples to be more explicit about
behaviour.