Replace the "no_std" Cargo feature with a default "std" feature.
Since Cargo will use the union of all requested features for a crate which is referenced multiple times, it is better to have features that enable things rather than disable them. Otherwise, the inclusion of an unrelated crate that doesn't rely on that feature can break a crate that does rely on that feature.