Make `Bus` memory usage more reasonable.
Previously a `Bus` would require a unique `VecDeque` for keeping pending
frames for each `Output`. It also did not provide a way of removing
outputs, meaning busses with a growing number of outputs would
infinitely grow in memory usage.

This new implementation only requires a single VecDeque for pending
frames, regardless of the number of outputs. Instead, we just store a
HashMap from the output's unique key to its current index into the
pending frames buffer.

Outputs now also implement Drop and will remove themselves from the
HashMap. It will also drain the front of the front of the buffer and
update the buffer indices for each of the outputs if the removed output
was the "latest" to receive frames.