Add `ring_buffer` module with `Fixed` and `Bounded` ring buffer types
`Fixed` is a ring buffer of fixed length, meaning each time an element
is `push`ed to the back, an element must also be popped from the front.

`Bounded` is a ring buffer with a fixed **max** length and allows for
pushing and popping to be performed separately.

Both types can be created from any type that implements the provided
`ring_buffer::Slice` and `SliceMut` traits. These traits are implemented
for all types that can provide a `slice` that can be used for pushing
and popping elements. Implementations are provided for fixed-size arrays
(0...128, 256, 512, 1024, 2048, 4096, 8192), `Vec`, `Box<[T]>` and
`&mut [T]`.

All public methods have documentation explaining behaviour and most have
examples that demonstrate use.

Some further tests have been added to the test module.

cc @andrewcsmith would be great to get your thoughts on this!

Closes #56.