Use `Fixed` ring buffer in `Sinc` interpolator rather than `VecDeque`
This should hopefully improve the efficiency of the Sinc interpolator
in some cases as it no longer requires a dynamically allocated queue.

`Fixed` is strictly more flexible in that it allows using stack,
allocated, owned and borrowed buffers.

It also slightly simplifies the implementation as the inner queue always
retains the same length following each call to `next_source_frame` -
using `Fixed` removes the need to manually pop elements from the front
of the `VecDeque`.

This change updates the `new` constructor to pass the desired ring
buffer rather than a `depth` and `padding` signal to add flexibility and
remove the need for dynamic allocation.

cc @andrewcsmith I updated and tested the `resample` example and it
seems to work identically :+1: Are you happy with this small API change?
Hopefully it should make `Sinc` strictly more flexible.