Add `Signal::is_exhausted` and `signal::lift`
Includes documentation with testable example for both.

Together, these address and close #86 - please see that issue to see my
thoughts and process behind the design.

Note that this is a very slight breaking change as
`signal::FromIterator` and `signal::FromInterleavedSamples` have been
updated to store the `next` frame in order to be able to behave
correctly and efficiently wrt exhaustiveness checking. This also means
that both signals now request the first frame from their respective
iterators upon construction, rather than when the first frame is
requested from the `Signal`.

@tuxmark5 you might be interested in this change and in particular the
new `signal::lift` function as it might allow you to get rid of the
`Rc<Cell<bool>>` you mentioned in #54.

@andrewcsmith as usual would be great to get your thoughts :) no problem
if you're busy!