Rename `sample` to `dasp`. Split features into multiple crates.
This PR is based on @PieterPenninckx's great work in #116 and some of
the past discussion that has occurred around improving the modularity of
the`sample` crate. Having the crate split into features as in #116 made
it easier for me to identify how to untangle the various modules into
independent crates.

**sample** -> **dasp**

The direction of splitting up the crate into multiple sub crates has
been on my mind for a long time, though the name `sample` always seemed
unsuitable to as an umbrella term for the sub-crates of the project.
After spending a bit of time throwing around different acronyms and
abbreviations, @JoshuaBatty suggested **dasp** for digital audio signal
processing. I think it rolls off the tongue pretty well!

**Repo structure**

The repository is now structured as a workspace consisting of a suite of
subcrates. The **dasp** crate is the top-level crate, re-exporting each
of the sub-crates and providing features that map not only to each of
the sub-crates, but also to each of the sub-crate features too.

**TODO**

I thought I'd post this PR now to allow for feedback and thoughts. That
said, there are still a few things I'd like to address before landing
this:

- [ ] Remove travis-ci in favour of github workflow CI with
  auto-publishing and a nice way of testing feature set permutations and
  with/without std.
- [ ] Port synth.rs and play_wav.rs examples from portaudio to cpal.
- [ ] Add CHANGELOG wich section on migrating from `sample` -> `dasp`.
- [ ] Complete the "crates" table within the top-level README.
- [ ] Add a graphviz png of the dasp dependency tree to the README.
- [ ] Add a note to the docs of feature-gated items specifying required
  features.
- [ ] Publish a new version of `sample` with a README notice about
  switching to `dasp`.

There are many other items I'd like to address, but they are likely best
left for future PRs.

cc @PieterPenninckx, @andrewcsmith

Closes #116.
Closes #112.