Improve flexibility of the **Window** trait
This allows for **Window** trait implementations to be more specific
about the supported phase and output amplitude types of the
implementation.

The Hanning and Rectangle implementations currently remain generic,
however we should consider restricting their implementations to floating
point sample types in order to not hide the intermediate conversions
that are required.